[No issue] Fix UTF-8 compilation
Per discussion here, the ydb_icu_version was actually never set
properly. This is because, using `env X=Y && env A=B` is syntatically
incorrect shell syntax. The correct syntax is `env X=Y A=B`.

https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1144#note_860166646

Also, when this was tested, it was found that we need to preserve
`ydb_icu_version` when running `sudo`.