[YottaDB/DB/YDB#850] Fix UTF-8 compilation on RHEL7 by using en_US.UTF-8 as a fallback locale if C.UTF-8 does not exist
A previous commit (6fc64b4) modified the Makefile to use the `C.UTF-8` locale to compile the YDBEncrypt plugin's M code.
This caused builds to fail on our in-house RHEL7 machine. After investigation, we discovered that the `C.UTF-8` locale
does not exist on RHEL7. It was added by Red Hat in version 8. Since RHEL7 is a supported operating system, the YDBEncrypt
install needs to work there. The fix for this is to use `en_US.UTF-8` as a fallback if the `C.UTF-8` locale is not found.