[YottaDB/DB/YDB#604] [V63013] Fix incorrect merge conflict resolutions in gtm_tls_impl.c
This commit fixes an incorrect merge conflict resolution in `gtm_tls_impl.c` which
caused numerous tests (mostly replication tests) to fail with errors like the below:

```
%YDB-E-TLSIOERROR, Error during TLS/SSL recv operation
%YDB-W-TLSHANDSHAKE, Connection to remote side using TLS/SSL protocol failed
```

The primary cause of this failure was an incorrect merge conflict resolution where
`tls_sock->flags` was only set to the new `GTMTLS_OP_NOSHUTDOWN` flag if the
`OPENSSL_VERSION_MAJOR` was less than 3. This caused tests involving encryption to fail
on machines where the OpenSSL version was version 3. A secondary cause was a line that
only ran if the `OPENSSL_VERSION_MAJOR` was 3 or more. In this line, the variable
`error_code` was not changed to `ssl_error_code` which is the equivalent of the previous
`error_code` but was left as `error_code` which is now a completely different new variable
that is an unsigned long rather than an int. This caused a few tests to fail if the machine
was running OpenSSL 3.  On machines running older versions of OpenSSL, encryption related
tests passed reliably.